feat: dynamic step attribute on price inputs
Fixed step="0.0001" was meaningless for exotic token pairs
(e.g. PEPE/WETH at ~0.000000001) and missing entirely on assumedPrice.

getPriceStep() computes step as 10^(floor(log10(value)) - 3),
keeping step ~0.1% of the value's magnitude across all price scales.

Applied to: minPrice, maxPrice, assumedPrice inputs in CalculatorInputs.

Renamed formatPriceInput.js -> priceInputUtils.js to reflect the file
now exports two price magnitude utilities.
